+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add conversion of standalone Vally `Trajectory` JSON objects to ATIF
+
 ## [0.12.1](https://github.com/waldekmastykarz/atifact/compare/v0.12.0...v0.12.1)
 
 ### Bug Fixes